healthcare-ai-samples/requirements.txt
azure-cognitiveservices-speech
openai
azure-ai-textanalytics
pandas
scikit-learn
streamlit
